Refactor configuration, database, and repository layers: - Introduce typed `Port` and `Env` with validation. - Rewrite PostgreSQL connection setup using `net/url` for safer DSN construction. - Migrate SQL queries to use templates for better maintainability. - Add improved error handling, logging, and default configuration values. - Refactor `docker-compose` to fix PostgreSQL volume mapping.

+++ internal/models/entity.go
@@ -3,6 +3,6 @@ package models
 // Entity defines the basic behavior for all storable models.
 // It allows repositories to manage entities generically while maintaining type safety.
 type Entity interface {
-	GetID() int
-	SetID(id int)
+	GetID() int64
+	SetID(id int64)
 }
